Added code formatting using pre-commit, black and pylint
