+++ Pipfile
@@ -10,6 +10,8 @@ uvicorn = "*"
 pydantic = "*"
 
 [dev-packages]
+pre-commit = "*"
+pylint = "*"
 
 [requires]
 python_version = "3.11"
