+++ main.py
@@ -1,19 +1,20 @@
+"""Entry point for the FastAPI application."""
 import random
 
 import chess
 from fastapi import FastAPI, Depends
 from fastapi.middleware.cors import CORSMiddleware
 
-from ChessBoardValidationMiddleware import ChessBoardValidationMiddleware, get_board
 from api_types import MoveResponse
+from middlewares import chess_board_validation_middleware, get_board
 from search_tree_algorithms import minimax
 
 app = FastAPI()
 
 # List of origins that are allowed to make requests to this API
 origins = [
-    'http://localhost',
-    'http://localhost:3000',
+    "http://localhost",
+    "http://localhost:3000",
 ]
 
 # Add CORS middleware to allow cross-origin requests
@@ -21,20 +22,20 @@ app.add_middleware(
     CORSMiddleware,
     allow_origins=origins,
     allow_credentials=True,
-    allow_methods=['*'],
-    allow_headers=['*'],
+    allow_methods=["*"],
+    allow_headers=["*"],
 )
 
 # Add board validation middleware to validate the chess board state
-app.middleware('http')(ChessBoardValidationMiddleware(app))
+app.middleware("http")(chess_board_validation_middleware)
 
 
-@app.get('/')
+@app.get("/")
 async def root():
     """
     Root endpoint that returns a simple greeting message.
     """
-    return 'Hello World'
+    return "Hello World"
 
 
 def log_move(board: chess.Board, move: chess.Move):
@@ -44,10 +45,10 @@ def log_move(board: chess.Board, move: chess.Move):
     :param board: A chess board object with the current game state.
     :param move: The chosen move as a chess.Move object.
     """
-    print(f'fen: {board.fen()}, move: {move.uci()}')
+    print(f"fen: {board.fen()}, move: {move.uci()}")
 
 
-@app.post('/api/random')
+@app.post("/api/random")
 async def random_ai(board: chess.Board = Depends(get_board)) -> MoveResponse:
     """
     API route for making a random move on the chessboard.
@@ -63,7 +64,7 @@ async def random_ai(board: chess.Board = Depends(get_board)) -> MoveResponse:
     return MoveResponse(fen=board.fen(), move=move.uci())
 
 
-@app.post('/api/minimax/{depth}')
+@app.post("/api/minimax/{depth}")
 async def minimax_ai(depth: int, board: chess.Board = Depends(get_board)) -> MoveResponse:
     """
     API route for making a move on the chessboard using the minimax algorithm
@@ -73,7 +74,7 @@ async def minimax_ai(depth: int, board: chess.Board = Depends(get_board)) -> Mov
     :param board: A chess board object with the current game state.
     :return: A MoveResponse object with the updated FEN and the chosen move.
     """
-    move, _ = minimax(board, depth, float('-inf'), float('inf'), True)
+    move, _ = minimax(board, depth, float("-inf"), float("inf"), True)
     best_move = chess.Move.from_uci(move)
     board.push(best_move)
